refactor: move dependencies debugger to guardrails file
