+++ mava/systems/jax/mappo/system.py
@@ -17,8 +17,10 @@
 from typing import Any, Tuple
 
 from mava.components.jax import building, executing, training, updating
-from mava.components.jax.building.debugging import ComponentDependencyDebugger
-from mava.components.jax.building.guardrails import ComponentDependencyGuardrails
+from mava.components.jax.building.guardrails import (
+    ComponentDependencyDebugger,
+    ComponentDependencyGuardrails,
+)
 from mava.specs import DesignSpec
 from mava.systems.jax import System
 from mava.systems.jax.mappo.components import ExtrasLogProbSpec
